Move bill runs controller into folder
We try and follow that the path you see in the url matches where the controller sits in the repo.
